LayoutTests/media/modern-media-controls/tracks-support/tracks-support-show-panel-fullscreen-expected.txt

Showing the tracks panel in fullscreen.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.querySelector('.tracks-panel').style.right is "217px"
PASS shadowRoot.querySelector('.tracks-panel').style.bottom is "101px"

PASS successfullyParsed is true

TEST COMPLETE
